Move AffineMatrixObject to the object module
This means that the AffineMatrix is now the only thing in the `affine` module, so I've moved the module level docs to the `AffineMatrix` struct and now `AffineMatrix` is in the `display` module
