+++ agb/src/display/affine.rs
@@ -1,86 +1,3 @@
-#![deny(missing_docs)]
-//! # Affine matrices for the Game Boy Advance
-//!
-//! An affine matrix represents an affine transformation, an affine
-//! transformation being one which preserves parallel lines (note that this
-//! therefore cannot represent perspective seen in games like Super Mario Kart).
-//! Affine matrices are used in two places on the GBA, for affine backgrounds
-//! and for affine objects.
-//!
-//! # Linear Algebra
-//! As a matrix, they can be manipulated using linear algebra. The short version
-//! of this section is to beware that the matrix is the inverse of the normal
-//! transformation matrices.
-//!
-//! One quick thing to point out at the start as it will become very relevant is
-//! that matrix-matrix multiplication is not commutative, meaning swapping the
-//! order changes the result, or **A** × **B** ≢ **B** × **A**. However,
-//! matrices are, at least in the case they are used here, associative, meaning
-//! (**AB**)**C** = **A**(**BC**).
-//!
-//! ## Normal (wrong on GBA!) transformation matrices
-//!
-//! As a start, normal transformation matrices will transform a shape from it's
-//! original position to it's new position. Generally when people talk about
-//! transformation matrices they are talking about them in this sense.
-//!
-//! > If **A** and **B** are transformation matrices, then matrix **C** = **A**
-//! > × **B** represents the transformation **A** performed on **B**, or
-//! > alternatively **C** is transformation **B** followed by transformation
-//! > **A**.
-//!
-//! This is not what they represent on the GBA! If you are looking up more
-//! information about transformation matrices bear this in mind.
-//!
-//! ## Correct (on GBA) transformation matrices
-//!
-//! On the GBA, the affine matrix works the other way around. The GBA wants to
-//! know for each pixel what colour it should render, to do this it applies the
-//! affine transformation matrix to the pixel it is rendering to lookup correct
-//! pixel in the texture.
-//!
-//! This describes the inverse of the previously given transformation matrices.
-//!
-//! Above I described the matrix **C** = **A** × **B**, but what the GBA wants
-//! is the inverse of **C**, or **C**<sup>-1</sup> = (**AB**)<sup>-1</sup> =
-//! **B**<sup>-1</sup> × **A**<sup>-1</sup>. This means that if we have the
-//! matrices **I** and **J** in the form the GBA expects then
-//!
-//! > Transformation **K** = **I** × **J** is the transformation **I** followed
-//! > by the transformation **J**.
-//!
-//! Beware if you are used to the other way around!
-//!
-//! ## Example, rotation around the centre
-//!
-//! To rotate something around its centre, you will need to move the thing such
-//! that the centre is at (0, 0) and then you can rotate it. After that you can
-//! move it where you actually want it.
-//!
-//! These can be done in the order I stated, **A** = **Move To Origin** ×
-//! **Rotate** × **Move to Final Position**. Or in code,
-//!
-//! ```rust,no_run
-//! # #![no_std]
-//! # #![no_main]
-//! use agb::fixnum::{Vector2D, Num, num};
-//! use agb::display::affine::AffineMatrix;
-//!
-//! # fn foo(_gba: &mut agb::Gba) {
-//! // size of our thing is 10 pixels by 10 pixels
-//! let size_of_thing: Vector2D<Num<i32, 8>> = (10, 10).into();
-//! // rotation by a quarter turn
-//! let rotation: Num<i32, 8> = num!(0.25);
-//! // the final position
-//! let position: Vector2D<Num<i32, 8>> = (100, 100).into();
-//!
-//! // now lets calculate the final transformation matrix!
-//! let a = AffineMatrix::from_translation(-size_of_thing / 2)
-//!     * AffineMatrix::from_rotation(rotation)
-//!     * AffineMatrix::from_translation(position);
-//! # }
-//! ```
-
 use core::ops::{Mul, MulAssign};
 
 use crate::fixnum::{FixedWidthSignedInteger, Num, SignedNumber, Vector2D, num, vec2};
@@ -93,6 +10,88 @@ use crate::fixnum::{FixedWidthSignedInteger, Num, SignedNumber, Vector2D, num, v
 /// a b x
 /// c d y
 /// 0 0 0
+/// ```
+///
+/// # Affine matrices for the Game Boy Advance
+///
+/// An affine matrix represents an affine transformation, an affine
+/// transformation being one which preserves parallel lines (note that this
+/// therefore cannot represent perspective seen in games like Super Mario Kart).
+/// Affine matrices are used in two places on the GBA, for affine backgrounds
+/// and for affine objects.
+///
+/// # Linear Algebra
+/// As a matrix, they can be manipulated using linear algebra. The short version
+/// of this section is to beware that the matrix is the inverse of the normal
+/// transformation matrices.
+///
+/// One quick thing to point out at the start as it will become very relevant is
+/// that matrix-matrix multiplication is not commutative, meaning swapping the
+/// order changes the result, or **A** × **B** ≢ **B** × **A**. However,
+/// matrices are, at least in the case they are used here, associative, meaning
+/// (**AB**)**C** = **A**(**BC**).
+///
+/// ## Normal (wrong on GBA!) transformation matrices
+///
+/// As a start, normal transformation matrices will transform a shape from it's
+/// original position to it's new position. Generally when people talk about
+/// transformation matrices they are talking about them in this sense.
+///
+/// > If **A** and **B** are transformation matrices, then matrix **C** = **A**
+/// > × **B** represents the transformation **A** performed on **B**, or
+/// > alternatively **C** is transformation **B** followed by transformation
+/// > **A**.
+///
+/// This is not what they represent on the GBA! If you are looking up more
+/// information about transformation matrices bear this in mind.
+///
+/// ## Correct (on GBA) transformation matrices
+///
+/// On the GBA, the affine matrix works the other way around. The GBA wants to
+/// know for each pixel what colour it should render, to do this it applies the
+/// affine transformation matrix to the pixel it is rendering to lookup correct
+/// pixel in the texture.
+///
+/// This describes the inverse of the previously given transformation matrices.
+///
+/// Above I described the matrix **C** = **A** × **B**, but what the GBA wants
+/// is the inverse of **C**, or **C**<sup>-1</sup> = (**AB**)<sup>-1</sup> =
+/// **B**<sup>-1</sup> × **A**<sup>-1</sup>. This means that if we have the
+/// matrices **I** and **J** in the form the GBA expects then
+///
+/// > Transformation **K** = **I** × **J** is the transformation **I** followed
+/// > by the transformation **J**.
+///
+/// Beware if you are used to the other way around!
+///
+/// ## Example, rotation around the centre
+///
+/// To rotate something around its centre, you will need to move the thing such
+/// that the centre is at (0, 0) and then you can rotate it. After that you can
+/// move it where you actually want it.
+///
+/// These can be done in the order I stated, **A** = **Move To Origin** ×
+/// **Rotate** × **Move to Final Position**. Or in code,
+///
+/// ```rust,no_run
+/// # #![no_std]
+/// # #![no_main]
+/// use agb::fixnum::{Vector2D, Num, num};
+/// use agb::display::AffineMatrix;
+///
+/// # fn foo(_gba: &mut agb::Gba) {
+/// // size of our thing is 10 pixels by 10 pixels
+/// let size_of_thing: Vector2D<Num<i32, 8>> = (10, 10).into();
+/// // rotation by a quarter turn
+/// let rotation: Num<i32, 8> = num!(0.25);
+/// // the final position
+/// let position: Vector2D<Num<i32, 8>> = (100, 100).into();
+///
+/// // now lets calculate the final transformation matrix!
+/// let a = AffineMatrix::from_translation(-size_of_thing / 2)
+///     * AffineMatrix::from_rotation(rotation)
+///     * AffineMatrix::from_translation(position);
+/// # }
 /// ```
 #[allow(missing_docs)]
 pub struct AffineMatrix<T> {
@@ -240,94 +239,6 @@ impl<T: SignedNumber> Mul for AffineMatrix<T> {
 impl<T: SignedNumber> MulAssign for AffineMatrix<T> {
     fn mul_assign(&mut self, rhs: Self) {
         *self = *self * rhs;
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-#[repr(C, packed(4))]
-/// An affine matrix that can be used in affine objects
-///
-/// ```txt
-/// a b
-/// c d
-/// ```
-#[allow(missing_docs)]
-pub struct AffineMatrixObject {
-    pub a: Num<i16, 8>,
-    pub b: Num<i16, 8>,
-    pub c: Num<i16, 8>,
-    pub d: Num<i16, 8>,
-}
-
-impl Default for AffineMatrixObject {
-    fn default() -> Self {
-        Self::from(AffineMatrix::<Num<i16, 8>>::identity())
-    }
-}
-
-impl<I, const N: usize> From<AffineMatrix<Num<I, N>>> for AffineMatrixObject
-where
-    I: FixedWidthSignedInteger,
-    i16: From<I>,
-{
-    fn from(value: AffineMatrix<Num<I, N>>) -> Self {
-        Self {
-            a: value.a.change_base(),
-            b: value.b.change_base(),
-            c: value.c.change_base(),
-            d: value.d.change_base(),
-        }
-    }
-}
-
-impl AffineMatrixObject {
-    #[must_use]
-    /// Converts to the affine matrix that is usable in performing efficient
-    /// calculations.
-    pub fn to_affine_matrix(&self) -> AffineMatrix<Num<i16, 8>> {
-        AffineMatrix {
-            a: self.a.change_base(),
-            b: self.b.change_base(),
-            c: self.c.change_base(),
-            d: self.d.change_base(),
-            x: 0.into(),
-            y: 0.into(),
-        }
-    }
-
-    #[must_use]
-    /// Converts from an affine matrix, wrapping if it overflows
-    pub fn from_affine_wrapping<I, const N: usize>(affine: AffineMatrix<Num<I, N>>) -> Self
-    where
-        I: FixedWidthSignedInteger,
-        i32: From<I>,
-    {
-        let a: Num<i32, 8> = affine.a.change_base();
-        let b: Num<i32, 8> = affine.b.change_base();
-        let c: Num<i32, 8> = affine.c.change_base();
-        let d: Num<i32, 8> = affine.d.change_base();
-
-        Self {
-            a: Num::from_raw(a.to_raw() as i16),
-            b: Num::from_raw(b.to_raw() as i16),
-            c: Num::from_raw(c.to_raw() as i16),
-            d: Num::from_raw(d.to_raw() as i16),
-        }
-    }
-
-    pub(crate) fn components(self) -> [u16; 4] {
-        [
-            self.a.to_raw() as u16,
-            self.b.to_raw() as u16,
-            self.c.to_raw() as u16,
-            self.d.to_raw() as u16,
-        ]
-    }
-}
-
-impl From<AffineMatrixObject> for AffineMatrix<Num<i16, 8>> {
-    fn from(mat: AffineMatrixObject) -> Self {
-        mat.to_affine_matrix()
     }
 }
